+++ src/main/java/UMC_8th/With_Run/common/config/properties/Constants.java
@@ -1,0 +1,6 @@
+package UMC_8th.With_Run.common.config.properties;
+
+public final class Constants {
+    public static final String AUTH_HEADER = "Authorization";
+    public static final String TOKEN_PREFIX = "Bearer ";
+}
