--- src/main/java/UMC_8th/With_Run/config/properties/Constants.java
@@ -1,6 +1,0 @@
-package UMC_8th.With_Run.config.properties;
-
-public final class Constants {
-    public static final String AUTH_HEADER = "Authorization";
-    public static final String TOKEN_PREFIX = "Bearer ";
-}
