--- src/main/java/UMC_8th/With_Run/config/properties/JwtProperties.java
@@ -1,22 +1,0 @@
-package UMC_8th.With_Run.config.properties;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.stereotype.Component;
-
-@Component
-@Getter
-@Setter
-@ConfigurationProperties("jwt.token")
-public class JwtProperties {
-    private String secretKey="";
-    private Expiration expiration;
-
-    @Getter
-    @Setter
-    public static class Expiration{
-        private Long access;
-        // TODO: refreshToken
-    }
-}
